Import toBeEq from bignumber-custom-matcher
