+++ packages/machine/test/integration/setup-then-set-state.spec.ts
@@ -9,16 +9,12 @@ import dotenv from "dotenv-safe";
 import { Contract, Wallet } from "ethers";
 import { AddressZero, WeiPerEther, Zero } from "ethers/constants";
 import { JsonRpcProvider } from "ethers/providers";
-import {
-  BigNumber,
-  hexlify,
-  Interface,
-  randomBytes,
-  SigningKey
-} from "ethers/utils";
+import { hexlify, Interface, randomBytes, SigningKey } from "ethers/utils";
 
 import { SetStateCommitment, SetupCommitment } from "../../src/ethereum";
 import { StateChannel } from "../../src/models";
+
+import { toBeEq } from "./bignumber-jest-matcher";
 
 // To be honest, 30000 is an arbitrary large number that has never failed
 // to reach the done() call in the test case, not intelligency chosen
@@ -40,23 +36,7 @@ let provider: JsonRpcProvider;
 let wallet: Wallet;
 let network: NetworkContext;
 
-// TODO: When we add a second use case of this custom mathcer,
-//       move it and its typigns into somewhere re-usable
-declare global {
-  namespace jest {
-    interface Matchers<R> {
-      toBeEq(expected: BigNumber): BigNumber;
-    }
-  }
-}
-expect.extend({
-  toBeEq(received: BigNumber, argument: BigNumber) {
-    return {
-      pass: received.eq(argument),
-      message: () => `expected ${received} not to be equal to ${argument}`
-    };
-  }
-});
+expect.extend({ toBeEq });
 
 // TODO: This will be re-used for all integration tests, so
 //       move it somewhere re-usable when we add a new test
